+++ src/functions/index.ts
@@ -1,0 +1,2 @@
+export { disability } from "./disability";
+export { dab } from "./dab";
